==> model_building_w_R.yang/05.logistic-regression.Rmd <==
---
title: "Logistic Regression"
output: 
  html_document:
    toc: true
    toc_depth: 4
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


## Read Data 
```{r}
dir0 = '/Users/mani/Dropbox/IIIBigDataClass_DataSync/_project/kkbox_churn_R/'

dir_train = paste(dir0, '/21a.train.field-select2_categ-transformed/', sep='')
dir_test = paste(dir0, '/21b.test.field-select2_categ-transformed/', sep='')

file = 'data01_decis-tree-selected_log-transf.csv'

df0_train = read.csv(paste(dir_train, file, sep=''), header=T, sep=",")
df0_test = read.csv(paste(dir_test, file, sep=''), header=T, sep=",")
```

```{r}
# fields handling / selection 
df0_train$X = NULL
df0_train$registered_via = as.factor(df0_train$registered_via)
df0_train$is_churn = NULL

df0_test$X = NULL
df0_test$registered_via = as.factor(df0_test$registered_via)
df0_test$is_churn = NULL
```

```{r}
str(df0_train)
```

## one-hot encoding
```{r}
library(dummies)
categ_columns = c('pay_method_mode', 'registered_via', 'auto_renew_mode', 'city')

df1_train = dummy.data.frame(df0_train, names=categ_columns)
df1_test = dummy.data.frame(df0_test, names=categ_columns)

str(df1_train)
```

## standardization
```{r}
df2_train = as.data.frame(scale(df1_train[,-40]))
df2_test = as.data.frame(scale(df1_test[,-40]))

df2_train = cbind(df2_train, df1_train['my_is_churn'])
df2_test = cbind(df2_test, df1_test['my_is_churn'])

str(df2_train)
```


## logistic regression
```{r}
fit = glm(my_is_churn~., 
          data=df2_train,
          family=binomial)
```

```{r}
summary(fit)
```


```{r results='hold'}
pred = predict(fit, newdata=df2_train, type='response')
pred = ifelse(pred > 0.5, 1, 0)
tb = table(df2_train$my_is_churn, pred)
accuracy = sum(diag(tb))/sum(tb)
sensitivity = tb[2,2]/(tb[2,1]+tb[2,2])
precision = tb[2,2]/(tb[1,2]+tb[2,2])
cat('on train set:\n')
print(tb)
cat('accuracy =', accuracy, '\n')
cat('sensitivity =', sensitivity, '\n')
cat('precision =', precision, '\n')
```

```{r results='hold'}
pred = predict(fit, newdata=df2_test, type='response')
pred = ifelse(pred > 0.5, 1, 0)
tb = table(df2_test$my_is_churn, pred)
accuracy = sum(diag(tb))/sum(tb)
sensitivity = tb[2,2]/(tb[2,1]+tb[2,2])
precision = tb[2,2]/(tb[1,2]+tb[2,2])
cat('on test set:\n')
print(tb)
cat('accuracy =', accuracy, '\n')
cat('sensitivity =', sensitivity, '\n')
cat('precision =', precision, '\n')
```

```{r results='hold'}
library(ROCR)
pred = predict(fit, newdata=df2_test, type='response')
predt = prediction(pred, df2_test$my_is_churn)
prf = performance(predt, measure="tpr", x.measure="fpr")
plot(prf)

auc = performance(predt, measure="auc")
auc = auc@y.values[[1]]
auc
```

### selected fields

```{r}
df2_train_sel = df2_train[,c(4,8,9,11,16,17,18,19,20,23,24,25,28,29,30,31,32,33,34,35,36,37,38,39)*(-1)]

str(df2_train_sel)
```

```{r}
fit = glm(my_is_churn~., 
          data=df2_train_sel,
          family=binomial)
```

```{r}
summary(fit)
```


```{r results='hold'}
pred = predict(fit, newdata=df2_train_sel, type='response')
pred = ifelse(pred > 0.5, 1, 0)
tb = table(df2_train_sel$my_is_churn, pred)
accuracy = sum(diag(tb))/sum(tb)
sensitivity = tb[2,2]/(tb[2,1]+tb[2,2])
precision = tb[2,2]/(tb[1,2]+tb[2,2])
cat('on train set:\n')
print(tb)
cat('accuracy =', accuracy, '\n')
cat('sensitivity =', sensitivity, '\n')
cat('precision =', precision, '\n')
```
